Update Rasbian version and move disk mounting and rebooting into if clauses

+++ setup_plex.sh
@@ -1,16 +1,13 @@
 #!/bin/bash
 # https://www.element14.com/community/community/raspberry-pi/raspberrypi_projects/blog/2016/03/11/a-more-powerful-plex-media-server-using-raspberry-pi-3
-VERSION=jessie
+VERSION=stretch #jessie
+
+# Disk mounting options
 LIBRARY=library
 DISK=${1:-sda1}
 FORMAT=${2:-ext4} #exfat #vfat ntfs
 
-echo "This will mount disk $DISK with the filesystem $FORMAT"
-read -p "Is this correct? " -n 1 -r
-#if [[ $REPLY =~ ^[Yy]$ ]]
-#then
-    # do dangerous stuff
-#fi
+
 
 sudo apt-get update && sudo apt-get upgrade -y
 sudo apt-get update && sudo apt-get dist-upgrade
@@ -55,23 +52,46 @@ sudo service plexmediaserver restart
 sudo usermod -aG pi plex
 sudo usermod -aG plex pi
 
-# Mount USB storage
-if [ "$FORMAT" == "ntfs" ]; then
-    sudo apt-get install ntfs-3g -y
-elif [ "$FORMAT" == "exfat" ]; then
-    sudo apt-get install exfat-fuse exfat-utils -y
+
+echo "This will mount disk $DISK with the filesystem $FORMAT"
+DEFAULT="Y"
+read -e -i "$DEFAULT" -n 1 -p "Is this correct? [Y/n]: " REPLY
+REPLY="${REPLY:-$DEFAULT}"
+if [[ $REPLY =~ ^[Yy]$ ]];
+then
+
+    # Install drivers if necessary
+    if [ "$FORMAT" == "ntfs" ]; then
+        sudo apt-get install ntfs-3g -y
+    elif [ "$FORMAT" == "exfat" ]; then
+        sudo apt-get install exfat-fuse exfat-utils -y
+    fi
+
+    # Edit fstab to automount at startup
+    # http://www.htpcguides.com/properly-mount-usb-storage-raspberry-pi/
+    sudo mkdir /mnt/$LIBRARY
+    echo "/dev/$DISK    /mnt/$LIBRARY   $FORMAT   defaults    0    0" | sudo tee -a /etc/fstab
+    # uid=pi,gid=pi,noatime
+
+    # Set permissions
+    sudo chown -R pi:pi /mnt/$LIBRARY
+    sudo chmod 775 -R /mnt/$LIBRARY
+    sudo setfacl -Rdm g:pi:rwx /mnt/$LIBRARY
+    sudo setfacl -Rm g:pi:rwx /mnt/$LIBRARY
+
+    # Mount USB storage
+    sudo mount -o uid=pi,gid=pi /dev/$DISK /mnt/$LIBRARY
+
+else
+    echo "Skipping disk mounting..."
 fi
-# http://www.htpcguides.com/properly-mount-usb-storage-raspberry-pi/
-sudo mkdir /mnt/$LIBRARY
-echo "/dev/$DISK    /mnt/$LIBRARY   $FORMAT   defaults    0    0" | sudo tee -a /etc/fstab
-# uid=pi,gid=pi,noatime
 
-sudo chown -R pi:pi /mnt/$LIBRARY
-sudo chmod 775 -R /mnt/$LIBRARY
-sudo setfacl -Rdm g:pi:rwx /mnt/$LIBRARY
-sudo setfacl -Rm g:pi:rwx /mnt/$LIBRARY
+echo "Plex server installation complete!"
 
-sudo mount -o uid=pi,gid=pi /dev/$DISK /mnt/$LIBRARY
-
-
-#sudo reboot
+read -p "Reboot now? [Y/n] " -n 1 -r REPLY
+if [[ $REPLY =~ ^[Yy]$ ]];
+then
+    sudo reboot
+else
+    echo "Please reboot manually. Exiting!"
+fi
